requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
openai==1.51.0
python-docx==1.1.2
PyPDF2==3.0.1
beautifulsoup4==4.12.3
jinja2==3.1.4
python-multipart==0.0.12
msal==1.31.0
chromadb>=0.5.0
openpyxl==3.1.5
python-pptx==1.0.2
opencv-python-headless==4.10.0.84
Pillow>=10.0.0
